gpxToExif/requirements.txt
piexif==1.1.3
Pillow==6.2.1